Only create jobs for derivations in jobsExpr
When using SKIP_ALREADY_BUILT we have to be careful to take the
intersection of the derivations from jobsExpr and the jobs that would be
built to realise those derivations.

Previously we weren't careful and we would create tons of jobs for
unbuilt paths in the dependency closure of those derivations.

This should fix it and only evaluate jobsExpr once (expensive).